// client/state/UI/ui.action-types.ts
export enum ActionTypes {
  TOGGLE_LOGIN = "TOGGLE_LOGIN",
  TOGGLE_REGISTER = "TOGGLE_REGISTER",
  OPEN_REGISTER = "OPEN_REGISTER",
  CLOSE_REGISTER = "CLOSE_REGISTER",
  OPEN_LOGIN = "OPEN_LOGIN",
  CLOSE_LOGIN = "CLOSE_LOGIN",
  TOGGLE_PASSWORD_RECOVER = "TOGGLE_PASSWORD_RECOVER",
  OPEN_PASSWORD_RECOVER = "OPEN_PASSWORD_RECOVER",
  CLOSE_PASSWORD_RECOVER = "CLOSE_PASSWORD_RECOVER",
  OPEN_RESET_PASSWORD = "OPEN_RESET_PASSWORD",
  CLOSE_RESET_PASSWORD = "CLOSE_RESET_PASSWORD",
  EMAIL_UPDATE_RECOVER = "EMAIL_UPDATE_RECOVER",
  TOGGLE_UNDER_CONSTRUCTION = "TOGGLE_UNDER_CONSTRUCTION",
  TOGGLE_UNDER_CONSTRUCTION_REQUEST = "TOGGLE_UNDER_CONSTRUCTION_REQUEST",
  TOGGLE_UNDER_CONSTRUCTION_SUCCESS = "TOGGLE_UNDER_CONSTRUCTION_SUCCESS",
  TOGGLE_UNDER_CONSTRUCTION_FAILURE = "TOGGLE_UNDER_CONSTRUCTION_FAILURE",
}